# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/Categoria/Create": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categorias"
                ],
                "summary": "Create a categoria",
                "parameters": [
                    {
                        "description": "Categoria data",
                        "name": "categoria",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.Categoria"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pkg.Body-types_Categoria"
                        }
                    }
                }
            }
        },
        "/api/Categoria/DeleteByID/{id}": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categorias"
                ],
                "summary": "Delete a categoria by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "categoria ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pkg.Body-string"
                        }
                    }
                }
            }
        },
...
        "/api/Categoria/GetByID/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categorias"
                ],
                "summary": "Get categorias",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Categoria ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pkg.Body-types_Categoria"
                        }
                    }
                }
            }
        },
        "/api/FormaPago/Create": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FormaPago"
                ],
                "summary": "Create a FormaPago",
                "parameters": [
                    {
                        "description": "FormaPago data",
                        "name": "FormaPago",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.FormaPago"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pkg.Body-types_FormaPago"
                        }
                    }
                }
            }
        },
        "/api/FormaPago/DeleteByID/{id}": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FormaPago"
                ],
                "summary": "Delete a FormaPago by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "FormaPago ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pkg.Body-string"
                        }
                    }
                }
            }
        },
...
        "/api/FormaPago/GetByID/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FormaPago"
                ],
                "summary": "Get FormaPago",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "FormaPago ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pkg.Body-types_FormaPago"
                        }
                    }
                }
            }
        },
        "/api/Product/Create": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Create a new product",
                "parameters": [
                    {
                        "description": "Product data",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.Producto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pkg.Body-types_Producto"
                        }
                    }
                }
            }
        },
        "/api/Product/DeleteByID/{id}": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Delete a product by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pkg.Body-string"
                        }
                    }
                }
            }
        },
...
        "/api/Product/GetByID/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Get a product by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pkg.Body-types_Producto"
                        }
                    }
                }
            }
        },
        "/api/Product/Update/{id}": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Update a product",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Product data",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.Producto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pkg.Body-types_Producto"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "pkg.Body-string": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string"
                }
            }
        },
        "pkg.Body-types_Categoria": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/types.Categoria"
                }
            }
        },
        "pkg.Body-types_FormaPago": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/types.FormaPago"
                }
            }
        },
        "pkg.Body-types_Producto": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/types.Producto"
                }
            }
        },
...
        "types.Categoria": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "nombre": {
                    "type": "string"
                }
            }
        },
        "types.FormaPago": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "nombre": {
                    "type": "string"
                }
            }
        },
        "types.Producto": {
            "type": "object",
            "properties": {
                "categoriaID": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "imagen": {
                    "type": "string"
                },
                "nombre": {
                    "type": "string"
                },
                "precio": {
                    "type": "number"
                },
                "stock": {
                    "type": "boolean"
                }
            }
        },
...
    }
}